deprecate model fit & sample methods
